Split compress.go by algorithm; rename LZW pools
Move LZW implementation (constants, pools, helpers, lzwCompress,
lzwDecompress) into lzw.go and snappy implementation
(maxPooledSnappyEncodeCap, snappyEncodeBufPool, putSnappyEncodeBuf,
snappyCompress, snappyDecompress) into snappy.go. compress.go retains
the algorithm-agnostic dispatch, the metric plumbing, and the
maxDecompressBytes ceiling shared by both algorithms.

Rename three pool symbols whose generic bytes-prefixed names hid their
LZW-only role: bytesBufferPool -> lzwScratchBufPool,
bytesReaderPool -> lzwSrcReaderPool,
emptyBytesReader -> lzwReaderInitSrc. Matches the existing helpers
(getLZWScratch / releaseLZWScratch) and the cap constant
(maxPooledLZWScratchCap), and aligns with the *BufPool convention used
by sibling pools (encodeBufPool, pushPullBufPool, snappyEncodeBufPool).

No behavioural change.
